LayoutTests/fast/dom/FileList-iterator-expected.txt


Tests that FileList objects have an iterator.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fileList.__proto__ is FileList.prototype
PASS Symbol.iterator in fileList is true
PASS forOfSucceeded is true
PASS 'entries' in fileList is false
PASS 'keys' in fileList is false
PASS 'forEach' in fileList is false
PASS 'values' in fileList is false
PASS successfullyParsed is true

TEST COMPLETE
